requirements.txt
flask==3.1.1
flask-session==0.8.0
flask-cors==5.0.1
gunicorn==23.0.0
python-dotenv==1.1.0
google-generativeai==0.8.5
groq==0.25.0
openai==1.82.0
bcrypt==4.3.0
stripe==12.2.0
libsql-experimental==0.0.52
httpx==0.28.1
requests==2.32.3
